s2: camera: fix compilation errors

+++ camera/QCamera2/HAL/QCamera2HWICallbacks.cpp
@@ -2372,7 +2372,7 @@ void QCamera2HardwareInterface::dumpJpegToFile(const void *data,
                     ssize_t written_len = write(file_fd, data, size);
                     fchmod(file_fd, S_IRUSR | S_IWUSR | S_IRGRP | S_IROTH);
                     CDBG_HIGH("%s: written number of bytes %ld\n",
-                            __func__, written_len);
+                            __func__, (long)written_len);
                     close(file_fd);
                 } else {
                     ALOGE("%s: fail t open file for image dumping", __func__);
@@ -2620,7 +2620,7 @@ void QCamera2HardwareInterface::dumpFrameToFile(QCameraStream *stream,
                         }
 
                         CDBG_HIGH("%s: written number of bytes %ld\n",
-                            __func__, written_len);
+                            __func__, (long)written_len);
                         close(file_fd);
                     } else {
                         ALOGE("%s: fail t open file for image dumping", __func__);
